input-modules/cleanup/requirements.txt
watchdog==0.8.2
python-dateutil==2.8.0